py-requirements.txt
markdown=3.4.3
python-frontmatter=1.0.0